Blufor Arsenal Updates
Added some more weapons, adjusted scopes, stripped out redundant backpacks, fixed KITE and MAXIKITE implementation, added equipment required for FW pilots. All scopes tested, should now be working ok.

Still seeing some things appearing in the arsenal that are not in the whitelist, and seeing things not in the whitelist that are not restricted - LRPAS ammo is currently not available.